requirements.txt
streamlit
langchain
langchain-google-genai
google-api-core